requirements.txt
cursor==1.3.4
mysql-connector==2.2.9
PyMySQL==1.0.2
python-dotenv==0.19.2
